+++ src/Framework/Bootstrap/Setup/GacelaConfigTransfer.php
@@ -21,6 +21,7 @@ use Gacela\Framework\Event\Dispatcher\ConfigurableEventDispatcher;
  * @psalm-import-type ServicesToExtendMap from ContainerConfigurationInterface
  * @psalm-import-type HandlerRegistriesMap from ContainerConfigurationInterface
  * @psalm-import-type TagsMap from ContainerConfigurationInterface
+ * @psalm-import-type AfterResolvingMap from ContainerConfigurationInterface
  * @psalm-import-type ContextualBindingsMap from ContainerConfigurationInterface
  * @psalm-import-type ConfigKeyValues from SetupGacelaInterface
  * @psalm-import-type SpecificListenersMap from ConfigurableEventDispatcher
@@ -44,6 +45,7 @@ final class GacelaConfigTransfer
      * @param HandlerRegistriesMap $handlerRegistries
      * @param ServiceFactoryMap $lazyServices
      * @param TagsMap $tags
+     * @param AfterResolvingMap $afterResolvingCallbacks
      */
     public function __construct(
         public readonly AppConfigBuilder $appConfigBuilder,
@@ -69,6 +71,7 @@ final class GacelaConfigTransfer
         public readonly array $handlerRegistries = [],
         public readonly array $lazyServices = [],
         public readonly array $tags = [],
+        public readonly array $afterResolvingCallbacks = [],
     ) {
     }
 }