+++ src/Framework/Bootstrap/Setup/Properties.php
@@ -24,6 +24,7 @@ use Gacela\Framework\Event\Dispatcher\EventDispatcherInterface;
  * @psalm-import-type ServicesToExtendMap from ContainerConfigurationInterface
  * @psalm-import-type HandlerRegistriesMap from ContainerConfigurationInterface
  * @psalm-import-type TagsMap from ContainerConfigurationInterface
+ * @psalm-import-type AfterResolvingMap from ContainerConfigurationInterface
  * @psalm-import-type ContextualBindingsMap from ContainerConfigurationInterface
  * @psalm-import-type ConfigKeyValues from SetupGacelaInterface
  * @psalm-import-type SpecificListenersMap from ConfigurableEventDispatcher
@@ -102,6 +103,9 @@ final class Properties
 
     /** @var ?TagsMap */
     public ?array $tags = null;
+
+    /** @var ?AfterResolvingMap */
+    public ?array $afterResolvingCallbacks = null;
 
     public function __construct()
     {